refactor: migrate shared components to generic directory and update imports
