+++ src/features/profile/components/ProfileContainer.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import Cover from '../../../shared/components/Cover';
-import Avatar from '../../../shared/components/Avatar';
+import Cover from '../../../components/generic/Cover';
+import Avatar from '../../../components/generic/Avatar';
 import ActionsPanel from './ActionsPanel';
 import UserInfo from './UserInfo';
 import Description from './Description';
